Removed changes out of scope. Moved function to utils.py, moved test to test_utils.py
