+++ bugbug/bug_features.py
@@ -11,34 +11,11 @@ from datetime import datetime, timedelta, timezone
 import dateutil.parser
 import pandas as pd
 from dateutil import parser
-from dateutil.rrule import DAILY, FR, MO, SA, SU, TH, TU, WE, rrule
 from libmozdata import versions
 from libmozdata.bugzilla import Bugzilla
 from sklearn.base import BaseEstimator, TransformerMixin
 
 from bugbug import bug_snapshot, repository
-
-
-def date_range(end_date, start_date):
-    return rrule(
-        DAILY, dtstart=start_date, until=end_date, byweekday=(MO, TU, WE, TH, FR)
-    )
-
-
-def business_day_range(end_date, start_date):
-    weekend_checker = [end_date]
-    if weekend_checker == list(rrule(DAILY, count=1, byweekday=(SA), dtstart=end_date)):
-        end_date = end_date + timedelta(days=2)
-    elif weekend_checker == list(
-        rrule(DAILY, count=1, byweekday=(SU), dtstart=end_date)
-    ):
-        end_date = end_date + timedelta(days=1)
-    if start_date > end_date:
-        range = date_range(start_date + timedelta(days=1), end_date)
-        return (len(list(range))) * -1.0
-    else:
-        range = date_range(end_date, start_date + timedelta(days=1))
-        return (len(list(range))) * 1.0
 
 
 def field(bug, field):
@@ -562,10 +539,10 @@ def get_time_to_fix(bug):
     if bug["cf_last_resolved"] is None:
         return None
 
-    return business_day_range(
-        dateutil.parser.parse(bug["cf_last_resolved"]),
-        dateutil.parser.parse(bug["creation_time"]),
-    )
+    return (
+        dateutil.parser.parse(bug["cf_last_resolved"])
+        - dateutil.parser.parse(bug["creation_time"])
+    ).total_seconds() / 86400
 
 
 class time_to_fix(single_bug_feature):
@@ -581,10 +558,10 @@ def get_time_to_assign(bug):
                 and change["removed"] in ("UNCONFIRMED", "NEW")
                 and change["added"] == "ASSIGNED"
             ):
-                return business_day_range(
-                    dateutil.parser.parse(history["when"]),
-                    dateutil.parser.parse(bug["creation_time"]),
-                )
+                return (
+                    dateutil.parser.parse(history["when"])
+                    - dateutil.parser.parse(bug["creation_time"])
+                ).total_seconds() / 86400
 
     return None
