+++ bugbug/utils.py
@@ -15,7 +15,7 @@ import tarfile
 import urllib.parse
 from collections import deque
 from contextlib import contextmanager
-from datetime import datetime
+from datetime import datetime, timedelta
 from functools import lru_cache
 from typing import Any, Iterator, Optional
 
@@ -28,6 +28,7 @@ import requests
 import scipy
 import taskcluster
 import zstandard
+from dateutil.rrule import DAILY, FR, MO, SA, SU, TH, TU, WE, rrule
 from pkg_resources import DistributionNotFound
 from requests.packages.urllib3.util.retry import Retry
 from sklearn.base import BaseEstimator, TransformerMixin
@@ -513,3 +514,26 @@ def escape_markdown(text: str) -> str:
         .replace(")", "\\)")
         .replace("|", "\\|")
     )
+
+
+# Remove business days from date range
+def date_range(end_date, start_date):
+    return rrule(
+        DAILY, dtstart=start_date, until=end_date, byweekday=(MO, TU, WE, TH, FR)
+    )
+
+
+def business_day_range(end_date, start_date):
+    weekend_checker = [end_date]
+    if weekend_checker == list(rrule(DAILY, count=1, byweekday=(SA), dtstart=end_date)):
+        end_date = end_date + timedelta(days=2)
+    elif weekend_checker == list(
+        rrule(DAILY, count=1, byweekday=(SU), dtstart=end_date)
+    ):
+        end_date = end_date + timedelta(days=1)
+    if start_date > end_date:
+        range = date_range(start_date + timedelta(days=1), end_date)
+        return (len(list(range))) * -1.0
+    else:
+        range = date_range(end_date, start_date + timedelta(days=1))
+        return (len(list(range))) * 1.0
