+++ scripts/generate_landings_risk_report.py
@@ -27,13 +27,13 @@ import markdown2
 import requests
 import sendgrid
 from dateutil.relativedelta import relativedelta
-from dateutil.rrule import DAILY, FR, MO, SA, SU, TH, TU, WE, rrule
 from tqdm import tqdm
 
 from bugbug import bug_features, bugzilla, db, phabricator, repository, test_scheduling
 from bugbug.models.bugtype import bug_to_types
 from bugbug.models.regressor import BUG_FIXING_COMMITS_DB, RegressorModel
 from bugbug.utils import (
+    business_day_range,
     download_check_etag,
     download_model,
     escape_markdown,
@@ -44,28 +44,6 @@ from bugbug.utils import (
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-
-def date_range(end_date, start_date):
-    return rrule(
-        DAILY, dtstart=start_date, until=end_date, byweekday=(MO, TU, WE, TH, FR)
-    )
-
-
-def business_day_range(end_date, start_date):
-    weekend_checker = [end_date]
-    if weekend_checker == list(rrule(DAILY, count=1, byweekday=(SA), dtstart=end_date)):
-        end_date = end_date + timedelta(days=2)
-    elif weekend_checker == list(
-        rrule(DAILY, count=1, byweekday=(SU), dtstart=end_date)
-    ):
-        end_date = end_date + timedelta(days=1)
-    if start_date > end_date:
-        range = date_range(start_date + timedelta(days=1), end_date)
-        return (len(list(range))) * -1.0
-    else:
-        range = date_range(end_date, start_date + timedelta(days=1))
-        return (len(list(range))) * 1.0
 
 
 TEST_INFOS_DB = "data/test_info.json"
